Fix serialization
